Fix some compiling warnings

+++ Visual Studio/Misc/LeetCode OJ/026. Remove Duplicates from Sorted Array/Solution.h
@@ -5,6 +5,6 @@ class Solution
 public:
     int removeDuplicates(int A[], int n)
     {
-        return unique(A, A + n) - A;
+        return static_cast<int>(unique(A, A + n) - A);
     }
 };
