+++ Visual Studio/Misc/LeetCode OJ/032. Longest Valid Parentheses/Solution.h
@@ -64,6 +64,6 @@ public:
             }
         }
 
-        return result;
+        return static_cast<int>(result);
     }
 };
