+++ Visual Studio/Misc/LeetCode OJ/209. Minimum Size Subarray Sum/Solution.h
@@ -33,6 +33,6 @@ public:
             }
         }
 
-        return minLength;
+        return static_cast<int>(minLength);
     }
 };
